Cập nhật giao diện cho comment bài viết, thêm animation chuyển trang và cải thiện trang tìm kiếm.

+++ app/src/main/java/com/example/uth_socials/ui/component/common/formatVND.kt
@@ -3,7 +3,7 @@ package com.example.uth_socials.ui.component.common
 import java.text.NumberFormat
 import java.util.Locale
 
-fun formatVND(price: Double): String {
+fun formatVND(price: Int): String {
     val locale = Locale("vi", "VN")
     val formatter = NumberFormat.getCurrencyInstance(locale)
     return formatter.format(price)
